Bug #1922680 [Coverity代码扫描][AP][sprdroid13_main][高]-File:[/bsp/modules/kernel5.4/npu/*]
[root cause  ]:coverity warning
[changes     ]:add init, add judgement of NULL pointer
[side effects]:no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:n6pro
[test case]:cnn_testbench
[reviewers]:wenming.wu

commit_template_version:v1

[change_type ] feature
[tag_product ] common

+++ vha/single/vha_cnn.c
@@ -206,7 +206,7 @@ static int do_cmd_cnn_submit(struct vha_cmd *cmd)
 				* (VHA_CR_OS(CNN_ALT_ADDRESS1)
 				- VHA_CR_OS(CNN_ALT_ADDRESS0));
 			/* record what alt address is in use */
-			alt_addrs_used |= 1 << user_cmd->regidx[i-1];
+			alt_addrs_used |= (uint64_t)1 << user_cmd->regidx[i-1];
 #if defined(HW_AX3)
 			/* Alternative addresses from 8 to 15 are
 			 * located in different place */
@@ -216,7 +216,7 @@ static int do_cmd_cnn_submit(struct vha_cmd *cmd)
 				* (VHA_CR_OS(CNN_ALT_ADDRESS1)
 				- VHA_CR_OS(CNN_ALT_ADDRESS0));
 			}
-			alt_addrs_used |= buf->req_type <<
+			alt_addrs_used |= (uint64_t)buf->req_type <<
 				(VHA_CR_OS(CNN_ALT_ADDRESS_USED_ALT_ADDR0_BUF_TYPE_SHIFT) +
 				user_cmd->regidx[i-1]);
 #elif defined(HW_AX2)
